test: `FakeAuthRepository` 클래스 제거 및 익명 객체로 대체
